requirements.txt
aiohttp
pillow
pyinstaller